+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- hnefatafl-client: show what move we are on.
+- End the game on time from the server.
 - hnefatafl-client: show captures with an 'X'.
 - hnefatafl-client: replace circles and triangle with chess pieces.
 - Remove archived games from the `ron` file.